errors.linux_kernel: improve parsing and output of specific errors
Make some specific parsers more robust so that they get more consistent
results, even if they are a bit less detailed.

Show the error location as part of the summary when it makes sense.

Custom output for specific kernel bug reports (spinlock bad magic) to
avoid over-specialization when it's not needed.

Signed-off-by: Ricardo Cañuelo <ricardo.canuelo@collabora.com>