Use hash signatures to uniquely identify state results
Extend the idea of error hash signatures and use it to identify the
parse results of states as well. Each state may define a set of
'signature fields' and, after the parsing is done and the data is
collected, these fields will be used to identify the result set and
generate a unique hash string for them.

These signatures aren't necessarily related to errors, they simply
represent a result.

For now, the new states '_summary', '_signature_fields' and '_signature'
are defined as hidden and are invisible unless the `--json-full` flag is
used.

This change also moves the hash generation to a common utility function.

Signed-off-by: Ricardo Cañuelo <ricardo.canuelo@collabora.com>